+++ api/controllers/BabyController.js
@@ -8,7 +8,8 @@ module.exports = {
   addBaby: function(req, res) {
     var babyVal = {
       name: req.body.name,
-      owner: req.body.owner
+      owner: req.body.owner,
+      gender: req.body.gender
     }
     BabyService.addBaby(babyVal, function(success) {
       res.json(success)
@@ -17,7 +18,8 @@ module.exports = {
   feedBaby: function(req, res) {
     var data = {
       babyId: req.body.baby,
-      when: req.body.when
+      when: req.body.when,
+      quantity: req.body.quantity
     }
     BabyService.feedBaby(data, function(success) {
       res.json(success)