+++ api/services/BabyService.js
@@ -12,13 +12,19 @@ module.exports = {
     })
   },
   feedBaby: function(data, next) {
-    Event.create({type: 'feed', date: new Date(data.when), baby: data.babyId}).exec(function(err, Event) {
+    Event.create({
+      type: 'feed',
+      date: new Date(data.when),
+      baby: data.babyId,
+      amount: data.quantity
+    }).exec(function(err, Event) {
       if(err) throw err;
       Baby.find(data.babyId)
         .populate("events")
         .exec(function (err, baby) {
           if(err) throw err;
           baby[0].lastfeed = new Date(data.when);
+          baby[0].lastfeedamount = data.quantity;
           baby[0].events.add(Event.id);
           baby[0].save(function(err, res) {
             console.log(res);