+++ assets/js/app.js
@@ -17,7 +17,21 @@ babyApp.controller('BabyCtrl',
   ['$scope', '$rootScope', '$interval', 'BabyService', 'EventService',
   function($scope, $rootScope, $interval, BabyService, EventService) {
   $scope.formData = {};
+  $scope.feedData = {};
   $scope.babies = [];
+  $scope.feedList = {
+    qty: null,
+    availableOptions: [
+      {id: '50', name: '50ml'},
+      {id: '60', name: '60ml'},
+      {id: '70', name: '70ml'},
+      {id: '80', name: '80ml'},
+      {id: '90', name: '90ml'},
+      {id: '100', name: '100ml'},
+      {id: '110', name: '110ml'},
+      {id: '120', name: '120ml'}
+    ]
+  };
 
   // Clock
   $scope.clock = Date.now();
@@ -34,6 +48,7 @@ babyApp.controller('BabyCtrl',
 
   // Button to add a baby
   $scope.addBaby = function() {
+    console.log($scope.formData);
     BabyService.addBaby($scope.formData).then(function(response) {
       $scope.babies.push(response);
       $scope.formData = {};
@@ -41,8 +56,8 @@ babyApp.controller('BabyCtrl',
   }
 
   // Button to add a feed event
-  $scope.addFeedEvent = function(babyId) {
-    EventService.feedBaby({when: $scope.clock, baby: babyId}).then(function(response) {
+  $scope.addFeedEvent = function(babyId, qty) {
+    EventService.feedBaby({when: $scope.clock, quantity: qty, baby: babyId}).then(function(response) {
       $scope.loadData();
     })
   }